Run the PulseAudio callbacks from the output
This is lots of code, but it's not particularly complex. Instead of
driving the callbacks from the input events, this drives them using the
output event.

For this to work, the input data delivered to the program are queued
into a lock-free ringbuffer. Then, when there is a need for output data,
PulseAudio calls its callback, and we put the data from the ring buffer
into a linear piece of memory, and call the callback as usual.

All the complexity and duplication is because we support f32 and i16
really.