Loopback mode implementation
While not an officially advertized feature in the PulseAudio API,
all PulseAudio "sink"-s have a "monitor mode" where in any of the
audio output devices there's a "loopback" input device which
replicates all audio for said output. This allows us to effectively
achieve recording any audio output as input within cubeb.

This commit is a continuation of said Draft shown in https://github.com/mozilla/cubeb/pull/691
but this time written in Rust. (yes it took me half a year but we're
here)

As I'm quite inexperienced when it comes to Rust (*especially so when it's
demanding code like FFI interfaces*) comments on the code or any
improvements that may be done are greatly appreciated.

Thanks.